fix type conflicts and prototypes

+++ ft_printf_pointer.c
@@ -6,7 +6,7 @@
 /*   By: jabenjam <jabenjam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/18 14:14:38 by jabenjam          #+#    #+#             */
-/*   Updated: 2020/02/19 11:49:27 by jabenjam         ###   ########.fr       */
+/*   Updated: 2020/02/19 12:08:38 by jabenjam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void		ft_pointer(unsigned long long pointer, t_data *f)
 	}
 	buffer = ft_ulltoa_base(pointer);
 	f->len = ft_strlen(buffer);
-	if ((size_t)f->precision < ft_strlen(buffer))
+	if (f->precision < ft_strlen(buffer))
 		f->precision = f->len;
 	if (f->left)
 		ft_padding_str(buffer, f);
